feat(follower): Register Twitter account by screennames
