+++ src/cli/follower.ts
@@ -1,57 +1,17 @@
 /* eslint-disable no-restricted-imports */
-import axios from 'axios';
 import prisma from '../prisma';
+import { twitterClient, UserObject } from './twitter-api';
 import { Prisma, SocialType } from '.prisma/client';
-
-const TWITTER_API_V2 = 'https://api.twitter.com/2/';
-const createTwitterClient = (bearerToken: string) => {
-  return async <T>(
-    endpoint: string,
-    { data, query }: { data?: Record<string, string>; query?: Record<string, string> },
-  ) => {
-    let url = `${TWITTER_API_V2}${endpoint}`;
-    if (query) {
-      const queryParams = Object.entries(query)
-        .map(([key, value]) => `${key}=${value}`)
-        .join('&');
-      url += '?' + queryParams;
-    }
-
-    console.log(`Requesting ${url}`);
-
-    const res = await axios.get<{ data: T; errors?: unknown[] }>(url, {
-      data,
-      headers: {
-        Authorization: `Bearer ${bearerToken}`,
-      },
-    });
-
-    return res.data;
-  };
-};
 
 const error = (message: unknown) => {
   console.error(message);
   process.exit(1);
 };
 
-type UserObject = {
-  id: string;
-  name: string;
-  public_metrics: {
-    followers_count: number;
-    following_count: number;
-    listed_count: number;
-    tweet_count: number;
-  };
-  username: string;
-};
-
 // Update Twitter follower table
 const fetch = async () => {
   // TODO: Add support for more than 100 users
   const today = new Date();
-  const twitterClient = createTwitterClient(process.env.TWITTER_BEARER_TOKEN);
 
   const fetchedAccounts = await prisma.socialFollower.findMany({ where: { date: today } });
   const accountsToFetch = await prisma.socialAccount.findMany({
@@ -101,19 +61,33 @@ const fetch = async () => {
 };
 
 // Register social account
-const register = async (socialType: SocialType, ids: bigint[]) => {
+const register = async (socialType: SocialType, ids: string[]) => {
   try {
+    // Normalize screen names (@***) to numerical id
+    const screenNamePattern = /^@/;
+    const screenNames = ids
+      .filter((id) => screenNamePattern.test(id))
+      .map((screenName) => screenName.replace('@', ''));
+    const numericalIds = ids.filter((id) => !screenNamePattern.test(id)).map(BigInt);
+
+    const { data: users } = await twitterClient<UserObject[]>('users/by', {
+      query: {
+        usernames: screenNames.join(','),
+      },
+    });
+    numericalIds.push(...users.map(({ id }) => BigInt(id)));
+
     // To avoid unique constraint error, first check existence before inserting.
     const existIds = (
       await prisma.socialAccount.findMany({
         select: { socialId: true },
         where: {
           socialType,
-          socialId: { in: ids },
+          socialId: { in: numericalIds },
         },
       })
     ).map(({ socialId }) => socialId);
-    const idsToInsert = ids.filter((id) => !existIds.includes(id));
+    const idsToInsert = numericalIds.filter((id) => !existIds.includes(id));
     if (idsToInsert.length === 0) {
       error('No new user to register.');
     } else if (existIds) {
@@ -156,7 +130,7 @@ const register = async (socialType: SocialType, ids: bigint[]) => {
         error(`No valid social type was specifide. Valid types are: ${Object.keys(socialTypes)}`);
       }
 
-      const ids = idsStr.split(',').map(BigInt);
+      const ids = idsStr.split(',');
       await register(socialType, ids);
 
       break;
